Remove email login and enable anonymous auth
