+++ __tests__/_app.test.tsx
@@ -5,7 +5,10 @@ function Dummy() {
   return <div data-testid="dummy" />;
 }
 
-test('renders page component', () => {
+test('renders nav and page component', () => {
   render(<App Component={Dummy} pageProps={{}} />);
+  expect(screen.getByText('Add Meal')).toBeInTheDocument();
+  expect(screen.getByText('History')).toBeInTheDocument();
+  expect(screen.getByText('Account')).toBeInTheDocument();
   expect(screen.getByTestId('dummy')).toBeInTheDocument();
 });
