+++ src/pages/account.tsx
@@ -1,0 +1,7 @@
+export default function Account() {
+  return (
+    <main>
+      <h1>Account</h1>
+    </main>
+  );
+}
