+++ src/pages/history.tsx
@@ -1,0 +1,7 @@
+export default function History() {
+  return (
+    <main>
+      <h1>History</h1>
+    </main>
+  );
+}
